refactor: Update Movie component
