+++ src/components/Movie.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect } from "react";
 import { useParams } from "react-router-dom";
 import { IMAGE_BASE_URL, POSTER_SIZE } from "../config";
-import { useMovieFetch } from "../hooks/useMoviesFetch";
 import NoImage from "../images/no_image.jpg";
 import Actor from "./Actor";
 import BreadCrumb from "./BreadCrumb";
@@ -9,6 +8,7 @@ import Grid from "./Grid";
 import MovieInfo from "./MovieInfo";
 import MovieInfoBar from "./MovieInfoBar";
 import Spinner from "./Spinner";
+import useMovieFetch from "../hooks/useMoviesFetch";
 
 /**
  * Renders the Movie component.
