+++ src/components/MovieInfoBar/props.ts
