+++ src/hooks/props.ts
@@ -1,9 +1,5 @@
-/**
- * Represents the cast of a movie.
- */
-export type Cast = {
-  character: string;
-  credit_id: string;
-  name: string;
-  profile_path: string;
-};
+import { Cast, Crew, MoviePropTypes } from "../Global.props";
+
+type MovieState = MoviePropTypes & { actors: Cast[]; directors: Crew[] };
+
+export type { MovieState };
